Class for object
Last week: implement an ObjectOnShapeDisplay class and create some
shape.

This week: Additional component of object class: center location (on
plane which is the pin coordinate), the center of mass of object,
vector (will be soon involved with rotation), and the plane (24x24array
replica of the object on real shape display).

The main focus is the operation of “add(obj)” and “subtract(obj)” which
will add and subtract the two object to form new shape and contain it
in the object you call these methods. In order to add and subtract,
“updatePlaneFromObject” and “updateObjectFromPlane” method are needed.
“clearWhiteArea” and “findCenter” are implemented to get center of mass
of the object update and ensure the accuracy of size of object by
deleting all white space around the object.