Add realtime typing display and fix upside down problem
Realtime typing display by function isTyping() let users see what
alphabet they type on keyboard and display that alphabet on the shape
display.

In the former work, when you apply ‘a.jpg’ to the shape display, it
seem not to be the correct side you would like to see. We should rotate
it.